android: Implement paired settings
Enables and disables editing on settings that rely on other boolean settings.
