+++ src/android/app/src/main/java/org/yuzu/yuzu_emu/model/SettingsViewModel.kt
@@ -24,6 +24,9 @@ class SettingsViewModel : ViewModel() {
     private val _shouldShowResetSettingsDialog = MutableLiveData(false)
     val shouldShowResetSettingsDialog: LiveData<Boolean> get() = _shouldShowResetSettingsDialog
 
+    private val _shouldReloadSettingsList = MutableLiveData(false)
+    val shouldReloadSettingsList: LiveData<Boolean> get() = _shouldReloadSettingsList
+
     fun setToolbarTitle(value: String) {
         _toolbarTitle.value = value
     }
@@ -38,6 +41,10 @@ class SettingsViewModel : ViewModel() {
 
     fun setShouldShowResetSettingsDialog(value: Boolean) {
         _shouldShowResetSettingsDialog.value = value
+    }
+
+    fun setShouldReloadSettingsList(value: Boolean) {
+        _shouldReloadSettingsList.value = value
     }
 
     fun clear() {
