+++ src/android/app/src/main/java/org/yuzu/yuzu_emu/utils/NativeConfig.kt
@@ -28,4 +28,6 @@ object NativeConfig {
     external fun getIsRuntimeModifiable(key: String): Boolean
 
     external fun getConfigHeader(category: Int): String
+
+    external fun getPairedSettingKey(key: String): String
 }
